Separate interactiontask files
